Extract `restore` tests from `sinon-test`.
Pulls out a suite into `test/util/core/restore-test` for better organisation.

As a side note, I am not sure that `sinon.restore` is a core util; should it infact be part of the public API and be moved to `lib/`?
