+++ test/sinon-test.js
@@ -73,39 +73,5 @@
             }
         },
 
-        ".restore": {
-            "restores all methods of supplied object": function () {
-                var methodA = function () {};
-                var methodB = function () {};
-                var obj = { methodA: methodA, methodB: methodB };
-
-                sinon.stub(obj);
-                sinon.restore(obj);
-
-                assert.same(obj.methodA, methodA);
-                assert.same(obj.methodB, methodB);
-            },
-
-            "only restores restorable methods": function () {
-                var stubbedMethod = function () {};
-                var vanillaMethod = function () {};
-                var obj = { stubbedMethod: stubbedMethod, vanillaMethod: vanillaMethod };
-
-                sinon.stub(obj, "stubbedMethod");
-                sinon.restore(obj);
-
-                assert.same(obj.stubbedMethod, stubbedMethod);
-            },
-
-            "restores a single stubbed method": function () {
-                var method = function () {};
-                var obj = { method: method };
-
-                sinon.stub(obj);
-                sinon.restore(obj.method);
-
-                assert.same(obj.method, method);
-            }
-        }
     });
 }(this));
